+++ config/routes.rb
@@ -3,4 +3,8 @@ RubyNative::Engine.routes.draw do
   namespace :push do
     resources :devices, only: :create
   end
+  namespace :auth do
+    get "start/:provider", to: "start#show", as: :start
+    resource :session, only: :show
+  end
 end