+++ lib/ruby_native.rb
@@ -1,6 +1,7 @@
 require "ruby_native/version"
 require "ruby_native/helper"
 require "ruby_native/native_detection"
+require "ruby_native/oauth_middleware"
 require "ruby_native/engine"
 
 module RubyNative
@@ -13,5 +14,6 @@ module RubyNative
     self.config = YAML.load_file(path).deep_symbolize_keys
     self.config[:app] ||= {}
     self.config[:app][:name] ||= "Ruby Native"
+    self.config[:auth] ||= {}
   end
 end